feat: cow module finished

+++ src/Controller/FarmController.php
@@ -2,30 +2,30 @@
 
 namespace App\Controller;
 
-use App\Entity\Cow;
 use App\Entity\Farm;
 use App\Form\FarmType;
+use App\Repository\FarmRepository;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Routing\Attribute\Route;
 
+#[Route('/farm')]
 class FarmController extends AbstractController
 {
-
-    #[Route('/farm', name: 'farm_index')]
-    public function index(EntityManagerInterface $entityManager, PaginatorInterface $paginator, Request $request)
+    #[Route('/', name: 'app_farm_index', methods: ['GET'])]
+    public function index(FarmRepository $farmRepository, EntityManagerInterface $entityManager, PaginatorInterface $paginator, Request $request): Response
     {
-        $farms = $entityManager->getRepository(Farm::class)->findAllFarm();
+        $farm = $farmRepository->findAllFarm();
 
-        $query = $entityManager->createQuery($farms);
+        $query = $entityManager->createQuery($farm);
 
         $pagination = $paginator->paginate(
             $query,
             $request->query->getInt('page', 1),
-            5
+            10
         );
 
         return $this->render('farm/index.html.twig', [
@@ -33,55 +33,66 @@ class FarmController extends AbstractController
         ]);
     }
 
-    #[Route('/farm/{id}/slaughter', name: 'farm_slaughter')]
-    public function SlaughterCows(EntityManagerInterface $entityManager, $id)
+    #[Route('/new', name: 'app_farm_new', methods: ['GET', 'POST'])]
+    public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
-        /** @var Cow[] $cows */
-        $cows = $entityManager->getRepository(Cow::class)->findAllCowsIdForSlaughter($id);
+        $farm = new Farm();
+        $form = $this->createForm(FarmType::class, $farm);
+        $form->handleRequest($request);
 
-        foreach ($cows as $cow) {
-            $cow = $entityManager->getRepository(Cow::class)->find($cow['id']);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $entityManager->persist($farm);
+            $entityManager->flush();
 
-            $cow->setSlaughtered(true);
+            $this->addFlash('success', 'Farm created successfully!');
+
+            return $this->redirectToRoute('app_farm_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        $entityManager->flush();
-
-        $this->addFlash('success', 'Cows have been slaughtered');
-
-        return $this->redirectToRoute('report_show', ['id' => $id]);
+        return $this->render('farm/new.html.twig', [
+            'farm' => $farm,
+            'form' => $form,
+        ]);
     }
 
-    #[Route('/farm/{id}', name: 'farm_edit')]
-    public function edit(EntityManagerInterface $entityManager, $id, Request $request)
+    #[Route('/{id}', name: 'app_farm_show', methods: ['GET'])]
+    public function show(Farm $farm): Response
     {
-        $form = $this->createForm(FarmType::class, $entityManager->getRepository(Farm::class)->find($id));
-        
+        return $this->render('farm/show.html.twig', [
+            'farm' => $farm,
+        ]);
+    }
+
+    #[Route('/{id}/edit', name: 'app_farm_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, Farm $farm, EntityManagerInterface $entityManager): Response
+    {
+        $form = $this->createForm(FarmType::class, $farm);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
 
-            $this->addFlash('success', 'Farm has been updated');
+            $this->addFlash('success', 'Farm updated successfully!');
 
-            return $this->redirectToRoute('farm_index');
+            return $this->redirectToRoute('app_farm_index', [], Response::HTTP_SEE_OTHER);
         }
 
-        return $this->render('farm/new.html.twig', [
-            'form' => $form->createView()
+        return $this->render('farm/edit.html.twig', [
+            'farm' => $farm,
+            'form' => $form,
         ]);
     }
 
-    #[Route('/farm/{id}', name: 'farm_delete', methods: ['POST'])]
-    public function delete(EntityManagerInterface $entityManager, $id)
+    #[Route('/{id}', name: 'app_farm_delete', methods: ['POST'])]
+    public function delete(Request $request, Farm $farm, EntityManagerInterface $entityManager): Response
     {
-        $farm = $entityManager->getRepository(Farm::class)->find($id);
+        if ($this->isCsrfTokenValid('delete'.$farm->getId(), $request->getPayload()->getString('_token'))) {
+            $entityManager->remove($farm);
+            $entityManager->flush();
 
-        $entityManager->remove($farm);
-        $entityManager->flush();
+            $this->addFlash('success', 'Farm deleted successfully!');
+        }
 
-        $this->addFlash('success', 'Farm has been deleted');
-
-        return $this->redirectToRoute('farm_index');
+        return $this->redirectToRoute('app_farm_index', [], Response::HTTP_SEE_OTHER);
     }
 }
